requirements.txt
Django==4.2.17
djangorestframework==3.15.2
python-decouple==3.8
django-model-utils==5.0.0
psycopg2==2.9.10
redis==5.2.0
celery==5.4.0
django-celery-beat==2.7.0
djangorestframework-simplejwt==5.3.1
